rename old inference algorithm
